+++ wp-content/themes/core/blocks/tribe/inline-notice/edit.js
@@ -8,15 +8,14 @@ import {
 import {
 	BaseControl,
 	Button,
+	Flex,
 	Modal,
 	PanelBody,
 	SelectControl,
 } from '@wordpress/components';
-import { useState } from '@wordpress/element';
-import IconPicker from 'components/IconPicker';
+import { RawHTML, useState } from '@wordpress/element';
+import IconPicker, { useRegisteredIcon } from 'components/IconPicker';
 import DynamicColorPicker from 'components/DynamicColorPicker';
-import { formatIconName } from 'blocks/tribe/icon-picker/utils';
-import { ICONS_LIST } from 'blocks/tribe/icon-picker/icons/icons-list';
 
 import './editor.pcss';
 
@@ -29,7 +28,6 @@ export default function Edit( { attributes, setAttributes, isSelected } ) {
 		iconPadding,
 		iconLabel,
 		iconSize,
-		searchQuery,
 		selectedIconColor,
 		selectedBgColor,
 		heading,
@@ -39,18 +37,7 @@ export default function Edit( { attributes, setAttributes, isSelected } ) {
 
 	const [ isModalOpen, setIsModalOpen ] = useState( false );
 
-	/**
-	 * Selected icon component based on the selectedIcon attribute
-	 */
-	const SelectedIconComponent =
-		ICONS_LIST.find( ( icon ) => icon.key === selectedIcon )?.component ||
-		null;
-
-	/**
-	 * Ensure selectedIcon is valid
-	 */
-	const validIcon =
-		ICONS_LIST.find( ( { key } ) => key === selectedIcon ) || null;
+	const validIcon = useRegisteredIcon( selectedIcon );
 
 	const classes = [
 		'b-inline-notice',
@@ -83,7 +70,6 @@ export default function Edit( { attributes, setAttributes, isSelected } ) {
 						iconPadding={ iconPadding }
 						iconLabel={ iconLabel }
 						iconSize={ iconSize }
-						searchQuery={ searchQuery }
 						selectedIconColor={ selectedIconColor }
 						selectedBgColor={ selectedBgColor }
 						onChange={ ( changed ) => setAttributes( changed ) }
@@ -109,42 +95,48 @@ export default function Edit( { attributes, setAttributes, isSelected } ) {
 						<BaseControl
 							__nextHasNoMarginBottom
 							id="icon-component"
-							className="controls-tribe-icon-picker icon-preview icon-card"
+							className="controls-tribe-icon-picker"
 						>
-							{ validIcon ? (
-								<>
-									<div
-										className="icon-image"
-										style={ {
-											backgroundColor:
-												selectedBgColor ||
-												'transparent',
-											color: selectedIconColor || 'white',
-											borderRadius: isRounded
-												? '50%'
-												: '0',
-										} }
-									>
-										<SelectedIconComponent
-											id="icon-component"
-											style={ {
-												color: selectedIconColor,
-											} }
-										/>
-									</div>
-									<p className="icon-name">
-										{ formatIconName( validIcon.name ) }
-									</p>
-								</>
-							) : (
-								__( 'No Icon Selected', 'tribe' )
-							) }
-							<Button
-								isPrimary
-								onClick={ () => setIsModalOpen( true ) }
+							<Flex
+								direction="column"
+								align="center"
+								gap={ 2 }
+								expanded={ false }
 							>
-								{ __( 'Open Icon Picker', 'tribe' ) }
-							</Button>
+								{ validIcon ? (
+									<>
+										<div
+											className="icon-image"
+											style={ {
+												backgroundColor:
+													selectedBgColor ||
+													'transparent',
+												color:
+													selectedIconColor ||
+													'white',
+												borderRadius: isRounded
+													? '50%'
+													: '0',
+											} }
+										>
+											<RawHTML>
+												{ validIcon.content }
+											</RawHTML>
+										</div>
+										<p className="icon-name">
+											{ validIcon.label }
+										</p>
+									</>
+								) : (
+									__( 'No Icon Selected', 'tribe' )
+								) }
+								<Button
+									isPrimary
+									onClick={ () => setIsModalOpen( true ) }
+								>
+									{ __( 'Open Icon Picker', 'tribe' ) }
+								</Button>
+							</Flex>
 						</BaseControl>
 						<DynamicColorPicker
 							colorAttribute="themeColor"
@@ -202,12 +194,7 @@ export default function Edit( { attributes, setAttributes, isSelected } ) {
 									padding: iconPadding + 'px',
 								} }
 							>
-								<SelectedIconComponent
-									id="icon-component"
-									style={ {
-										color: selectedIconColor,
-									} }
-								/>
+								<RawHTML>{ validIcon.content }</RawHTML>
 							</div>
 						</>
 					) : (