+++ CHANGELOG.md
@@ -4,10 +4,15 @@ All notable changes to this project will be documented in this file. The format 
 on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). Each changelog entry gets prefixed with the category of the
 item (Added, Changed, Depreciated, Removed, Fixed, Security).
 
-@@ [2026.08]
-- CHore WordPress Core update to 7.0.3, plugin, composer & npm package updates.
+## [2026.08]
+
+- Chore: WordPress Core update to 7.0.3, plugin, composer & npm package updates.
+- Removed: Tribe Icon Picker block
+- Added: Support for core Icon block; Modern Tribe custom icon collection.
+- Updated: Icon Card & Inline Notice blocks now pull icons from icon collections (no public editor UI available for Icon picking currently).
 
 ## [2026.06]
+
 - Chore: WordPress Core update to v7.0, plugins, Composer & NPM deps updates. update ESLint config file format.
 
 ## [2026.05]